Naos 2.0/requirements.txt
openai
pyyaml
flask